Trois commits sur une branch, ca fait un gorille

+++ src/Controller/ContentController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Content;
+use App\Entity\Users;
 use DateTime;
 use DateTimeZone;
 use Symfony\Component\HttpFoundation\RedirectResponse;
@@ -11,6 +12,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Twig\Environment;
 use Twig\Error\LoaderError;
 use Twig\Error\RuntimeError;
@@ -39,8 +41,13 @@ class ContentController extends AbstractController
      */
     public function index(): Response
     {
-        $contents = $this->getDoctrine()->getRepository(Content::class)->findAll();
+//        $contents = $this->getDoctrine()->getRepository(Content::class)->findAll();
+//        $contents = array_reverse($contents);
+
+        $contents = $this->getDoctrine()->getRepository(Content::class)->findAllWithUsers();
         $contents = array_reverse($contents);
+//        dd($contents);
+
 
         return new Response($this->twig->render('home.html.twig', [
             'contents' => $contents
@@ -63,6 +70,17 @@ class ContentController extends AbstractController
         $newContent = $request->get('content');
         $parent = $request->get('parent');
         $content = new Content();
+
+        if ($request->files->get('image') != null){
+            /** @var UploadedFile $uploadedFile */
+            $uploadedFile = $request->files->get('image');
+            $originalFilename = pathinfo($uploadedFile->getClientOriginalName(), PATHINFO_FILENAME);
+            $newFilename = $originalFilename.'-'.uniqid().'.'.$uploadedFile->guessExtension();
+            $destination = $this->getParameter('kernel.project_dir').'/public/uploads';
+            $uploadedFile->move($destination, $newFilename);
+            $content->setPicture($newFilename);
+        }
+
 
 
 //        dd($parent);
@@ -124,6 +142,7 @@ class ContentController extends AbstractController
             $contents = $this->getDoctrine()->getRepository(Content::class)->findChildren($idContent);
             $post = $this->getDoctrine()->getRepository(Content::class)->find($idContent);
             $entityManager = $this->getDoctrine()->getManager();
+            $this->denyAccessUnlessGranted('edit', $post);
             $entityManager->remove($post);
             foreach ($contents as $content) {
                 $entityManager->remove($content);
@@ -131,12 +150,14 @@ class ContentController extends AbstractController
             $entityManager->flush();
         } else {
             $post = $this->getDoctrine()->getRepository(Content::class)->find($idContent);
+            $this->denyAccessUnlessGranted('edit', $post);
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($post);
             $entityManager->flush();
         }
 
 
+        /** @var TYPE_NAME $contents */
         $errors = $validator->validate($contents);
         if (count($errors) > 0) {
             return new Response((string)$errors, 400);
